+++ serialize/SerializeElement.h
@@ -29,9 +29,11 @@ public:
 	}
 	SerializeElement() : LinkedElement<SerializeIF>(this) {
 	}
+
 	T entry;
-	ReturnValue_t serialize(uint8_t** buffer, uint32_t* size,
-			const uint32_t max_size, bool bigEndian) const {
+
+	ReturnValue_t serialize(uint8_t** buffer, size_t* size,
+			const size_t max_size, bool bigEndian) const {
 		return SerializeAdapter<T>::serialize(&entry, buffer, size, max_size, bigEndian);
 	}
